Now effects thumbnails preview Stable save/load picture on Pause/Resume (error correction no 1) Faces: try to show Landmarks. Adjusting photo in ImageView Implements methods for drawing 3D-shapes from Empty3-library-3d and Transparency. PixM char -> double (?) FIX#1

+++ app/src/main/java/one/empty3/feature20220726/PixM.java
@@ -50,7 +50,7 @@ public class PixM extends MBitmap {
 
     public PixM(Bitmap image) {
         super(image.getWidth(), image.getHeight());
-        int [] colorComponents = new int[4];
+        double [] colorComponents = new double[4];
         for (int i = 0; i < image.getWidth(); i++) {
             for (int j = 0; j < image.getHeight(); j++) {
                 int rgb = image.getPixel(i, j);
@@ -65,7 +65,7 @@ public class PixM extends MBitmap {
 
     public PixM(Bitmap image, boolean isBitmap) {
         super(image);
-        int[] colorComponents = new int[4];
+        double[] colorComponents = new double[4];
         for (int i = 0; i < image.getWidth(); i++) {
             for (int j = 0; j < image.getHeight(); j++) {
                 int rgb = image.getPixel(i, j);
